Перенесены функции вкладки 3 в отдельные модули
